requirements.txt
Flask==2.0.2
joblib==1.1.0
matplotlib==3.5.0
matplotlib-inline==0.1.2
numpy==1.21.2
pandas==1.3.4
PyWavelets==1.2.0
scikit-learn==1.0.1
scipy==1.7.3
seaborn==0.11.2
